requirements.txt
asttokens==2.2.1
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
et-xmlfile==1.1.0
executing==1.2.0
Flask==2.2.2
icecream==2.1.3
idna==3.4
imageio==2.23.0
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
networkx==2.8.8
numpy==1.23.5
openai==0.25.0
openpyxl==3.0.10
pandas==1.5.2
pandas-stubs==1.5.2.221124
Pillow==9.3.0
pygame==2.1.2
Pygments==2.13.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
simple-term-menu==1.5.2
six==1.16.0
soupsieve==2.3.2.post1
terminal-copilot==1.0.7
tqdm==4.64.1
types-pytz==2022.6.0.1
typing_extensions==4.4.0
urllib3==1.26.13
Werkzeug==2.2.2
wikipedia==1.4.0
zipp==3.11.0
